+++ product_portfolio/models.py
@@ -301,6 +301,10 @@ class Product(BaseProductMixin, FieldChangesMixin, KeywordsMixin, DataspacedMode
     def get_improve_packages_from_purldb_url(self):
         return self.get_url("improve_packages_from_purldb")
 
+    @property
+    def cyclonedx_bom_ref(self):
+        return str(self.uuid)
+
     def can_be_changed_by(self, user):
         perms = guardian.shortcuts.get_perms(user, self)
         has_change_permission_on_product = "change_product" in perms
@@ -508,9 +512,15 @@ class Product(BaseProductMixin, FieldChangesMixin, KeywordsMixin, DataspacedMode
         """Fetch and update the vulnerabilties of all the Package of this Product."""
         return fetch_for_queryset(self.all_packages, self.dataspace)
 
-    def get_vulnerability_qs(self):
+    def get_vulnerability_qs(self, prefetch_related_packages=False):
         """Return a QuerySet of all Vulnerability instances related to this product"""
-        return Vulnerability.objects.filter(affected_packages__in=self.packages.all())
+        qs = Vulnerability.objects.filter(affected_packages__in=self.packages.all())
+
+        if prefetch_related_packages:
+            package_qs = Package.objects.filter(product=self).only_rendering_fields()
+            qs = qs.prefetch_related(models.Prefetch("affected_packages", package_qs))
+
+        return qs
 
 
 class ProductRelationStatus(BaseStatusMixin, DataspacedModel):